Solved the issue with space input.
In the previous version, the password could not have spaces in it. This problem was posed because of the exit methods defined with READ in FoxPro. This version solves the problem by checking last key entered in READ, if the last key was a space then a space is appended to the pswd variable.